requirements.txt
progressbar2
numpy
pytest
nltk